Create LoginRegisterForm component and use it through loginPage, so it can be access through the page or as modal
